Move StWidgetPrivate to st-private.h so it can be used in st-label.c
